+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - none
 
 # v2025.2.1-beta
+- jslint - Update shell-function shDirHttplinkValidate() with pragma <\!!--novalidate--\>.
 - ci - Upgrade cpplint.py used in ci to v2.0.0.
 - sqlite - Rename custom-sql-function MEDIAN() to MEDIAN2().
 - sqlite - Enable builtin-sql-functions MEDIAN(), PERCENTILE(), PERCENTILE_CONT(), PERCENTILE_DISC().